Move components to shared + new Universe page + refactor
