+++ web/shared/lib/components/Typewriter.svelte
